refactor: break circular imports in shopify cart and webgl canvas
Both cycles had the same shape — a React context defined in the
provider/index file, a child component importing that context back, and the
index importing the child. Extracted each context into its own leaf module so
the back-edge disappears:

- lib/integrations/shopify/cart/cart-store-context.ts (CartContext +
  useCartContext + types), imported by both cart-context.tsx and modal/index.tsx
- lib/webgl/components/canvas/canvas-context.ts (CanvasContext), imported by
  both canvas/index.tsx and webgl.tsx

Public API preserved: cart-context.tsx and canvas/index.tsx re-export the moved
symbols, so existing import paths (add-to-cart, show-cart) are unchanged.
Behavior-preserving; no runtime changes.
