+++ lib/integrations/shopify/cart/modal/index.tsx
@@ -7,7 +7,7 @@ import { createContext, use, useState } from 'react'
 import { Image } from '@/components/ui/image'
 import { Link } from '@/components/ui/link'
 import { removeItem, updateItemQuantity } from '../actions'
-import { useCartContext } from '../cart-context'
+import { useCartContext } from '../cart-store-context'
 import s from './modal.module.css'
 
 interface ModalContextType {
