+++ lib/webgl/components/canvas/canvas-context.ts
@@ -1,0 +1,11 @@
+'use client'
+
+import { createContext } from 'react'
+import type tunnel from 'tunnel-rat'
+
+export type CanvasContextValue = {
+  WebGLTunnel?: ReturnType<typeof tunnel>
+  DOMTunnel?: ReturnType<typeof tunnel>
+}
+
+export const CanvasContext = createContext<CanvasContextValue>({})
